Split app into components
